requirements.txt
Flask==3.0.3
PyPDF2==3.0.1
openai==1.43.0
python-dotenv==1.0.1
gunicorn==23.0.0
httpx==1.0.0b0
hubspot-api-client==9.0.0
pytest==8.3.2
requests==2.32.3
tqdm==4.66.5
pydantic_core==2.23.3
pydantic==2.9.1
